+++ src/frontend/apps/impress/src/features/header/components/LaGaufre.tsx
@@ -2,6 +2,13 @@ import { Gaufre } from '@gouvfr-lasuite/integration';
 import '@gouvfr-lasuite/integration/dist/css/gaufre.css';
 import Script from 'next/script';
 import React from 'react';
+import { createGlobalStyle } from 'styled-components';
+
+const GaufreStyle = createGlobalStyle`
+  .lasuite-gaufre-btn{
+    box-shadow: inset 0 0 0 0 !important;
+  }
+`;
 
 export const LaGaufre = () => (
   <>
@@ -10,6 +17,7 @@ export const LaGaufre = () => (
       strategy="lazyOnload"
       id="lasuite-gaufre-script"
     />
-    <Gaufre />
+    <GaufreStyle />
+    <Gaufre variant="small" />
   </>
 );