+++ src/frontend/apps/impress/src/layouts/index.ts
@@ -1,2 +1,1 @@
 export * from './MainLayout';
-export * from './PageLayout';