+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 
 - 💄(frontend) error alert closeable on editor #284
 - ♻️(backend) Change email content #283
+- 💄(frontend) integrate new ui #286
 
 ## Fixed